+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next Up
+# 6.2.3
 
 - Weapons and armour did not link correctly in imported journals (possible DDB change, or maybe long standing issue).
 - The DDB Importer Activity Macro feature name has been tweaked, and some of the settings made clearer.